# scripts/sdk/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nocodb",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/db/meta/projects/{projectId}/info": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "projectId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Project info",
        "operationId": "project-meta-get",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Node": {
                      "type": "string"
                    },
                    "Arch": {
                      "type": "string"
                    },
                    "Platform": {
                      "type": "string"
                    },
                    "Docker": {
                      "type": "boolean"
                    },
                    "Database": {
                      "type": "string"
                    },
                    "ProjectOnRootDB": {
                      "type": "string"
                    },
                    "RootDB": {
                      "type": "string"
                    },
                    "PackageVersion": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Project"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "multipleOf": 1
            },
            "in": "query"
          }
        ]
      }
    },
...
    "/api/v1/db/meta/projects/": {
      "parameters": [],
      "get": {
        "summary": "Project list",
        "operationId": "project-list",
        "responses": {
          "201": {
            "$ref": "#/components/responses/ProjectList"
          }
        },
        "description": "Read project details",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "page"
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "pageSize"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "sort"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "xc-auth",
            "description": "Auth token"
          }
        ],
        "tags": [
          "Project"
        ]
      },
      "post": {
        "summary": "Project create",
        "operationId": "project-create",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Project"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "external": {
                        "type": "boolean",
                        "default": false
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "tags": [
          "Project"
        ]
      }
    },
    "/api/v1/db/meta/projects/{projectId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "projectId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Project read",
        "operationId": "project-read",
        "description": "Read project details",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "xc-auth",
            "description": "Auth token"
          }
        ],
        "tags": [
          "Project"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Project delete",
        "operationId": "project-delete",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Project"
        ]
      },
      "patch": {
        "summary": "Project update",
        "operationId": "project-update",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Project"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        }
      }
    },
...
    "/api/v1/db/meta/projects/{projectId}/cost": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "projectId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Project compare cost",
        "operationId": "project-cost",
        "description": "Project compare cost",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "xc-auth",
            "description": "Auth token"
          }
        ],
        "tags": [
          "Project"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/db/meta/projects/{projectId}/has-empty-or-null-filters": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "projectId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "",
        "operationId": "project-has-empty-or-null-filters",
        "parameters": [],
        "tags": [
          "Project"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Project": {
        "title": "Project",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "meta": {
            "type": [
              "string",
              "object"
            ]
          },
          "color": {
            "type": "string"
          },
          "deleted": {
            "type": [
              "string",
              "boolean"
            ]
          },
          "order": {
            "type": "number"
          },
          "bases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Base"
            }
          },
          "is_meta": {
            "type": "boolean"
          },
          "prefix": {
            "type": "string"
          },
          "created_at": {},
          "updated_at": {},
          "slug": {
            "type": "string"
          }
        }
      },
      "ProjectList": {
        "description": "",
        "type": "object",
        "x-examples": {
          "example-1": {
            "bases": {
              "list": [
                {
                  "id": "string",
                  "project_id": "string",
                  "alias": "string",
                  "host": "string",
                  "port": 0,
                  "username": "string",
                  "password": "string",
                  "database": "string",
                  "url": "string",
                  "params": "string",
                  "type": "string",
                  "ssl": "string"
                }
              ],
              "pageInfo": {
                "pageSize": 0,
                "totalRows": 0,
                "sort": "string",
                "isFirstPage": true,
                "isLastPage": true
              }
            }
          }
        },
        "properties": {
          "list": {
            "type": "array",
            "uniqueItems": true,
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "pageInfo": {
            "$ref": "#/components/schemas/Paginated"
          }
        }
      },
      "Base": {
        "title": "Base",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "alias": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "is_meta": {
            "type": "boolean",
            "readOnly": true
          },
          "config": {},
          "created_at": {},
          "updated_at": {},
          "inflection_column": {
            "type": "string"
          },
          "inflection_table": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
...
      "Sort": {
        "title": "Sort",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fk_model_id": {
            "type": "string"
          },
          "fk_column_id": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "project_id": {
            "type": "string",
            "readOnly": true
          },
          "base_id": {
            "type": "string",
            "readOnly": true
          }
        }
      },
...
      "Paginated": {
        "title": "Paginated",
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "integer"
          },
          "totalRows": {
            "type": "integer"
          },
          "sort": {
            "type": [
              "string",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/Sort"
            }
          },
          "isFirstPage": {
            "type": "boolean"
          },
          "isLastPage": {
            "type": "boolean"
          },
          "page": {
            "type": "number"
          }
        }
      },
...
    },
...
    "responses": {
      "ProjectList": {
        "description": "Example response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProjectList"
            },
            "examples": {
              "example-1": {
                "value": {
                  "projects": {
                    "list": [
                      {
                        "id": "string",
                        "title": "string",
                        "status": "string",
                        "description": "string",
                        "meta": "string",
                        "color": "string",
                        "deleted": "string",
                        "order": 0,
                        "bases": [
                          {
                            "id": "string",
                            "project_id": "string",
                            "alias": "string",
                            "type": "string",
                            "is_meta": true,
                            "config": null,
                            "created_at": null,
                            "updated_at": null
                          }
                        ],
                        "is_meta": true,
                        "prefix": "string",
                        "created_at": null,
                        "updated_at": null
                      }
                    ],
                    "pageInfo": {
                      "pageSize": 0,
                      "totalRows": 0,
                      "sort": "string",
                      "isFirstPage": true,
                      "isLastPage": true,
                      "page": 0
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "xc-auth": {
        "name": "API Key",
        "type": "apiKey",
        "in": "header",
        "description": "JWT access token"
      },
      "xc-shared-base-id": {
        "name": "API Key",
        "type": "apiKey",
        "in": "header",
        "description": "Shared base uuid"
      }
    }
  }
}